move numerics impls directly into winrt
